+++ system/scripts/combat.js
@@ -59,7 +59,7 @@ export class CombatL5r5e extends Combat {
 
             if (combatant.actor.data.type === "npc" && combatant.actor.data.data.type === "minion") {
                 // Minion NPCs can generate initiative value without a check, using their focus or vigilance attribute
-                initiative = cfg.prepared.minion ? data.focus : data.vigilance;
+                initiative = cfg.prepared.minion ? data.focus : data.is_compromised ? 1 : data.vigilance;
             } else {
                 // PC and Adversary
                 const isPc = combatant.actor.data.type === "character";
@@ -69,7 +69,7 @@ export class CombatL5r5e extends Combat {
                 if (isPrepared === "null") {
                     isPrepared = data.prepared;
                 }
-                initiative = isPrepared ? data.focus : data.vigilance;
+                initiative = isPrepared ? data.focus : data.is_compromised ? 1 : data.vigilance;
 
                 // Roll formula
                 if (!formula) {