+++ system/scripts/actors/base-sheet.js
@@ -159,7 +159,7 @@ export class BaseSheetL5r5e extends ActorSheet {
             event.stopPropagation();
             const elmt = $(event.currentTarget).data("toggle");
             const tgt = html.find("." + elmt);
-            tgt.hasClass("toggle-active") ? tgt.removeClass("toggle-active") : tgt.addClass("toggle-active");
+            tgt.toggleClass("toggle-active");
         });
 
         // *** Everything below here is only needed if the sheet is editable ***