+++ system/scripts/actors/twenty-questions-dialog.js
@@ -165,7 +165,7 @@ export class TwentyQuestionsDialog extends FormApplication {
             event.stopPropagation();
             const elmt = $(event.currentTarget).data("toggle");
             const tgt = html.find("." + elmt);
-            tgt.hasClass("toggle-active") ? tgt.removeClass("toggle-active") : tgt.addClass("toggle-active");
+            tgt.toggleClass("toggle-active");
         });
 
         // BT Next